+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 
 - Configurable parallel jobs when running `qft evaluate-compression`
-- Progress bar to `qft evaluate-compression`
+- Multi and single (configurable) Progress bar to `qft evaluate-compression`
 
 
 ## 0.9.0 - 2024-07-13